refactor: change recruitment api
