+++ packages/gdsc-dju-admin/src/pages/Admin/AdminApplicants/index.tsx
@@ -2,7 +2,7 @@ import { useAtom } from 'jotai';
 import React, { useEffect } from 'react';
 import { AdminSectionWrapper } from './styled';
 import { useRecoilValue } from 'recoil';
-import { recruitmentAtom } from '../../../atoms/recruitmentAtom';
+import { recruitmentAtom } from '../../../store/recruitmentAtom';
 import API from '../../../apis/index';
 import { useSearchParams } from 'react-router-dom';
 import { useLocation } from 'react-router';
