+++ packages/gdsc-dju-admin/src/pages/Admin/AdminEmail/index.tsx
@@ -3,10 +3,10 @@ import { addDoc, collection } from 'firebase/firestore';
 import { AnimatePresence } from 'framer-motion';
 import { useAtom } from 'jotai';
 import React, { useEffect, useState } from 'react';
-import { alertAtom } from '../../../atoms/alertAtom';
-import { loaderAtom } from '../../../atoms/loaderAtom';
-import { modalAtom } from '../../../atoms/modalAtom';
-import { userAtom } from '../../../atoms/userAtom';
+import { alertAtom } from '../../../store/alertAtom';
+import { loaderAtom } from '../../../store/loaderAtom';
+import { modalAtom } from '../../../store/modalAtom';
+import { userAtom } from '../../../store/userAtom';
 import { db } from '../../../firebase/firebase';
 import { GDSCButton } from '../../../components/common/Button';
 import CheckBoxCard from '../../../components/common/CheckBoxCard';
