--- packages/gdsc-dju-web/src/apis/hooks/useGetMemberData.ts
@@ -1,18 +1,0 @@
-import useSWR from 'swr';
-import Api from '../index';
-
-async function getMemberNickname() {
-  const res = await Api.getMemberNickname();
-  return res.data;
-}
-
-export function useGetMemberNickname() {
-  const { data } = useSWR(
-    [`/api/member/onBoarding/nickname`],
-    getMemberNickname,
-    { suspense: true },
-  );
-  return {
-    data: data && data.data,
-  };
-}
