--- packages/gdsc-dju-web/src/types/admin.d.ts
@@ -1,5 +1,0 @@
-export interface memberWarning {
-  content: string;
-  title: string;
-  toUser: string;
-}
